api/requirements.txt
flask
flask-cors
mediapipe
opencv-python-headless
numpy
python-dotenv
openai
scikit-learn